+++ templates/workspaces/base/terraform/variables.tf
@@ -103,6 +103,9 @@ variable "client_secret" {
   default     = ""
   description = "The client secret of the workspace in the identity provider, this is passed in so that we may return it as an output."
 }
+variable "ui_client_id" {
+  type = string
+}
 variable "sp_id" {
   type        = string
   default     = ""