+++ templates/workspaces/base/terraform/workspace.tf
@@ -36,6 +36,10 @@ module "aad" {
   workspace_owner_object_id      = var.workspace_owner_object_id
   aad_redirect_uris_b64          = var.aad_redirect_uris_b64
   create_aad_groups              = var.create_aad_groups
+  ui_client_id                   = var.ui_client_id
+  auth_client_id                 = var.auth_client_id
+  auth_client_secret             = var.auth_client_secret
+  auth_tenant_id                 = var.auth_tenant_id
 
   depends_on = [
     azurerm_role_assignment.keyvault_deployer_ws_role,